requirements.txt
pandas
python-dateutil
sqlalchemy
requests
pymysql
